pyproject.toml
[project]
name = "catapult-2026-best-ever-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.3.3",
    "opencv-python>=4.12.0.88",
    "pyside6>=6.10.0",
    "python-dotenv>=1.2.2",
    "spotipy>=2.26.0",
    "tqdm>=4.67.1",
]
